Use single copy of compressed snmp.yml
