+++ pkg/integrations/snmp_exporter/snmp_exporter.go
@@ -11,13 +11,9 @@ import (
 	"github.com/go-kit/log"
 	"github.com/grafana/agent/pkg/integrations"
 	"github.com/grafana/agent/pkg/integrations/config"
+	snmp_common "github.com/grafana/agent/pkg/integrations/snmp_exporter/common"
 	snmp_config "github.com/prometheus/snmp_exporter/config"
-	"gopkg.in/yaml.v2"
 )
-
-//go:generate curl https://raw.githubusercontent.com/prometheus/snmp_exporter/v0.20.0/snmp.yml --output snmp.yml
-//go:embed snmp.yml
-var content []byte
 
 // DefaultConfig holds the default settings for the snmp_exporter integration.
 var DefaultConfig = Config{
@@ -68,16 +64,6 @@ func init() {
 	integrations.RegisterIntegration(&Config{})
 }
 
-// LoadEmbeddedConfig loads the SNMP config via a file using the go:embed directive.
-func LoadEmbeddedConfig() (*snmp_config.Config, error) {
-	cfg := &snmp_config.Config{}
-	err := yaml.UnmarshalStrict(content, cfg)
-	if err != nil {
-		return nil, err
-	}
-	return cfg, nil
-}
-
 // New creates a new snmp_exporter integration
 func New(log log.Logger, c *Config) (integrations.Integration, error) {
 	var modules *snmp_config.Config
@@ -88,7 +74,7 @@ func New(log log.Logger, c *Config) (integrations.Integration, error) {
 			return nil, fmt.Errorf("failed to load snmp config from file %v: %w", c.SnmpConfigFile, err)
 		}
 	} else {
-		modules, err = LoadEmbeddedConfig()
+		modules, err = snmp_common.LoadEmbeddedConfig()
 		if err != nil {
 			return nil, fmt.Errorf("failed to load embedded snmp config: %w", err)
 		}
